Pages files cleanup. Admin dashboad component cleanup
