+++ src/app/components/admin-dashboard/accounts/add-account/add-account.component.tsx
@@ -1,35 +1,8 @@
 import React from 'react';
 import { useHistory } from 'react-router';
-
-import {
-    TextField,
-    Button,
-    Input,
-    InputLabel,
-    Typography,
-    Select,
-    MenuItem,
-} from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
+import { TextField, Button, InputLabel, Typography, Select, MenuItem } from '@material-ui/core';
 import { useForm, useAPI } from '../../../../hooks';
-
-const useStyles = makeStyles(() => ({
-    form: {
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'space-evenly',
-        height: '35vh',
-        width: '100%',
-        maxWidth: '80ch',
-    },
-    inputContainer: {
-        'width': '100%',
-        '& > *': {
-            width: '100%',
-        },
-    },
-}));
-
+import { useStyles } from './add-account-component-styles';
 const AddAccount: React.FC = () => {
     const classes = useStyles({});
