+++ src/app/components/admin-dashboard/accounts/list/list-component-styles.tsx
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/styles';
+export const useStyles = makeStyles(() => ({
+    header: {
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+    },
+    loading: {
+        marginLeft: '1ch',
+    },
+}));
