+++ src/app/pages/admin-dashboard/admin-dashboard-page/admin-dashboard-page-styles.tsx
@@ -1,0 +1,21 @@
+import { makeStyles } from '@material-ui/core/styles';
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        display: 'flex',
+    },
+    appBar: {
+        width: `calc(100% - 240px)`,
+        marginLeft: 240,
+    },
+    spacer: {
+        flex: '1',
+    },
+    main: {
+        display: 'flex',
+        flexDirection: 'column',
+        flexGrow: 1,
+        backgroundColor: theme.palette.background.default,
+        padding: theme.spacing(3),
+    },
+    toolbar: theme.mixins.toolbar,
+}));
