+++ src/app/pages/admin-dashboard/cohort/manage/manage.page.tsx
@@ -2,24 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { ListCohorts } from '../../../../components';
 import { Button, Typography } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(() => ({
-    root: {
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'space-evenly',
-        width: '100%',
-    },
-    header: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-    },
-    loading: {
-        marginLeft: '1ch',
-    },
-}));
+import { useStyles } from './manage-page-styles';
 
 const ManageCohortPage: React.FC = () => {
     const classes = useStyles();
