+++ src/app/pages/index.ts
@@ -12,7 +12,7 @@ export * from './dashboard/child/edit/edit.page';
 export * from './child-dashboard/child-dashboard.page';
 export * from './pdf-display/pdf-display.page';
 
-export * from './admin-dashboard/admin-dashboard.page';
+export * from './admin-dashboard/admin-dashboard-page/admin-dashboard.page';
 export * from './admin-onboarding';
 export * from './error-page/error-page';
 export * from './help-page/help-page';
